Added some insurance so that I don't lose my mind in the future.
Please, just give me a timestamp. The absolute circus gymnastics I have to do in order to get an accurate figure might as well secure me a career at one by the time I'm finished.
Why provide timestamps for everything else, but not referrals and popular content?

W h y    There is no reason. None. It's pure sadism. UTC midnight passes, the old numbers go out, the new ones come in, and I have no clue which is which because I'm left with nothing but a sum total of the last 14 days.

I'm gonna have to combine the view data and try to deduce it, and if that fails, an estimate, an inaccurate figure, is all that GitHub has provided me with.

                  w h y

w h y
                         w h y