Refactored reporter.rs and report_model.rs in to report.rs
I'm still keeping the other two files that this new one will replace for now. I decided to first just deal with views and clones before spending an eternity trying to get accurate metrics for referrals and popular content, and get nothing done. The referrals and popular content are a bonus to come.